Refactor: FirestoreLetterManager 분리(Write, Get)
