+++ Kabinett/Data/DefaultUseCases/DefaultImportLetterUseCase.swift
@@ -13,13 +13,13 @@ final class DefaultImportLetterUseCase {
     private let logger: Logger
     private let authManager: AuthManager
     private let writerManager: FirestoreWriterManager
-    private let letterManager: FirestoreLetterManager
+    private let letterManager: FirestoreLetterWriteManager
     private let letterStorageManager: FirestorageLetterManager
     
     init(
         authManager: AuthManager,
         writerManager: FirestoreWriterManager,
-        letterManager: FirestoreLetterManager,
+        letterManager: FirestoreLetterWriteManager,
         letterStorageManager: FirestorageLetterManager
     ) {
         self.logger = Logger(
