+++ Kabinett/Data/DefaultUseCases/DefaultWriteLetterUseCase.swift
@@ -13,13 +13,13 @@ final class DefaultWriteLetterUseCase {
     private let logger: Logger
     private let authManager: AuthManager
     private let writerManager: FirestoreWriterManager
-    private let letterManager: FirestoreLetterManager
+    private let letterManager: FirestoreLetterWriteManager
     private let letterStorageManager: FirestorageLetterManager
     
     init(
         authManager: AuthManager,
         writerManager: FirestoreWriterManager,
-        letterManager: FirestoreLetterManager,
+        letterManager: FirestoreLetterWriteManager,
         letterStorageManager: FirestorageLetterManager
     ) {
         self.logger = Logger(
@@ -34,7 +34,7 @@ final class DefaultWriteLetterUseCase {
 }
 
 extension DefaultWriteLetterUseCase: WriteLetterUseCase {
-    func saveLetter(font: String, 
+    func saveLetter(font: String,
                     postScript: String?,
                     envelope: String,
                     stamp: String,
