+++ Kabinett/Data/Services/FirestoreLetterWriteManager.swift
@@ -1,0 +1,159 @@
+//
+//  FirestoreLetterWriteManager.swift
+//  Kabinett
+//
+//  Created by JIHYE SEOK on 12/5/24.
+//
+
+import Foundation
+import FirebaseFirestore
+import os
+
+enum LetterSaveError: Error {
+    case failedToSaveToMe
+    case failedToSaveSent
+    case failedToSaveReceived
+    case failedToSaveBoth
+    case bothUsersNotFound
+    case invalidLetterId
+}
+
+struct Query {
+    let key: String
+    let value: String
+}
+
+final class FirestoreLetterWriteManager {
+    private let logger: Logger
+    private let db = Firestore.firestore()
+    
+    init() {
+        self.logger = Logger(
+            subsystem: "co.kr.codegrove.Kabinett",
+            category: "FirestoreLetterWriteManager"
+        )
+    }
+    
+    // MARK: - Firestore Letter 저장
+    func saveLetterToFireStore(
+        letter: Letter,
+        fromUserId: String?,
+        toUserId: String?
+    ) async -> Result<Bool, any Error> {
+        do {
+            let fromUserDoc = fromUserId.flatMap { !$0.isEmpty ? db.collection("Writers").document($0) : nil }
+            let toUserDoc = toUserId.flatMap { !$0.isEmpty ? db.collection("Writers").document($0) : nil }
+            
+            let fromUserSnapshot = fromUserDoc != nil ? try await fromUserDoc!.getDocument() : nil
+            let toUserSnapshot = toUserDoc != nil ? try await toUserDoc!.getDocument() : nil
+            
+            // User 검색용 필드 추가
+            let searchUser: [String] = [
+                letter.fromUserName,
+                String(letter.fromUserKabinettNumber ?? 0),
+                letter.toUserName,
+                String(letter.toUserKabinettNumber ?? 0)
+            ].compactMap { $0.lowercased() }
+                .flatMap { $0.split(separator: " ").map(String.init) }
+            
+            // fromUser가 존재하고, fromUserId와 toUserId가 같은 경우 -> ToMe
+            if let fromUserSnapshot = fromUserSnapshot, fromUserSnapshot.exists && fromUserId == toUserId {
+                do {
+                    var letterData = try Firestore.Encoder().encode(letter)
+                    letterData["searchUser"] = searchUser
+                    
+                    try await fromUserDoc!.collection("ToMe").addDocument(data: letterData)
+                    return .success(true)
+                } catch {
+                    logger.error("Save ToMe Error: \(error.localizedDescription)")
+                    return .failure(LetterSaveError.failedToSaveToMe)
+                }
+                // fromUser, toUser가 존재하는데 두 User가 다른 경우 -> Sent, Received
+            } else if let fromUserSnapshot = fromUserSnapshot, let toUserSnapshot = toUserSnapshot,
+                      fromUserSnapshot.exists && toUserSnapshot.exists && fromUserId != toUserId {
+                var sentSaveError: Error?
+                var receivedSaveError: Error?
+                
+                do {
+                    var sentLetter = letter
+                    sentLetter.isRead = true
+                    
+                    var letterSentData = try Firestore.Encoder().encode(sentLetter)
+                    letterSentData["searchUser"] = searchUser
+                    
+                    try await fromUserDoc!.collection("Sent").addDocument(data: letterSentData)
+                } catch {
+                    sentSaveError = error
+                }
+                
+                do {
+                    var letterData = try Firestore.Encoder().encode(letter)
+                    letterData["searchUser"] = searchUser
+                    
+                    try await toUserDoc!.collection("Received").addDocument(data: letterData)
+                } catch {
+                    receivedSaveError = error
+                }
+                
+                if let _ = sentSaveError, let _ = receivedSaveError {
+                    logger.error("Save Sent, Received Error: \(LetterSaveError.failedToSaveBoth)")
+                    return .failure(LetterSaveError.failedToSaveBoth)
+                } else if let _ = sentSaveError {
+                    logger.error("Save Sent Error: \(LetterSaveError.failedToSaveSent)")
+                    return .failure(LetterSaveError.failedToSaveSent)
+                } else if let _ = receivedSaveError {
+                    logger.error("Save Received Error: \(LetterSaveError.failedToSaveReceived)")
+                    return .failure(LetterSaveError.failedToSaveReceived)
+                }
+                return .success(true)
+                // fromUser가 존재하고, toUser가 존재하지 않을 때 -> Sent
+            } else if let fromUserSnapshot = fromUserSnapshot, fromUserSnapshot.exists,
+                      toUserSnapshot == nil || !toUserSnapshot!.exists {
+                do {
+                    var sentLetter = letter
+                    sentLetter.isRead = true
+                    
+                    var letterSentData = try Firestore.Encoder().encode(sentLetter)
+                    letterSentData["searchUser"] = searchUser
+                    
+                    try await fromUserDoc!.collection("Sent").addDocument(data: letterSentData)
+                    return .success(true)
+                } catch {
+                    logger.error("Save Sent Error: \(error.localizedDescription)")
+                    return .failure(LetterSaveError.failedToSaveSent)
+                }
+                // fromUser가 존재하지 않고, toUser가 존재할 때 -> Received
+            } else if fromUserSnapshot == nil || !fromUserSnapshot!.exists,
+                      let toUserSnapshot = toUserSnapshot, toUserSnapshot.exists {
+                do {
+                    var letterData = try Firestore.Encoder().encode(letter)
+                    letterData["searchUser"] = searchUser
+                    
+                    try await toUserDoc!.collection("Received").addDocument(data: letterData)
+                    return .success(true)
+                } catch {
+                    logger.error("Save Received Error: \(error.localizedDescription)")
+                    return .failure(LetterSaveError.failedToSaveReceived)
+                }
+                // 두 User가 모두 없을 때 -> failure
+            } else {
+                logger.error("Both User Found Error: \(LetterSaveError.bothUsersNotFound)")
+                return .failure(LetterSaveError.bothUsersNotFound)
+            }
+        } catch {
+            logger.error("Save Firestore Error: \(error.localizedDescription)")
+            return .failure(error)
+        }
+    }
+    
+    func findDocuments<T: Codable>(
+        by query: Query,
+        as type: T.Type
+    ) async throws -> [T] {
+        return try await db.collection("Writers")
+            .whereField(query.key, isEqualTo: query.value)
+            .getDocuments()
+            .documents
+            .map { try $0.data(as: type) }
+    }
+}
